prometheus_service_docker/requirements.txt
fastapi
uvicorn
numpy
requests
termcolor
pyyaml